\name{coerceFLBRP}
\docType{methods}
\alias{coerce,FLBRP,FLSR-method}
\alias{coerce,FLBRP,FLStock-method}
\title{Coercion from FLBRP to other classes}
\description{
}
\section{Generic function}{
methodName(class1, class2) 
}
\section{Methods}{\describe{
\item{signature(obj1=class1,obj2=class2):}{}
}}
\author{The FLR Team}
\seealso{\link[originalPkg]{methodName}}
\examples{
}
\keyword{methods}
