\name{propagate}
\docType{methods}
\alias{propagate,FLlst,list-method}
\title{extend FLBRP objects along the iter dimension}
\description{
}
\section{Generic function}{
 propagate(object, iter) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
