\name{setPlusGroupFLBRP}
\docType{methods}
\alias{setPlusGroup,FLBRP-method}
\title{setPlusGroup method for FLBRP}
\description{
}
\section{Generic function}{
methodName(class1, class2) 
}
\section{Methods}{\describe{
\item{signature(obj1=class1,obj2=class2):}{}
}}
\author{The FLR Team}
\seealso{\link[originalPkg]{methodName}}
\examples{
}
\keyword{methods}
